+++ packages/swc-plugin-workflow/transform/tests/errors/invalid-exports/output-client.js
@@ -1,4 +1,3 @@
-import { runStep as __private_run_step } from "workflow/api";
 /**__internal_workflows{"steps":{"input.js":{"validStep":{"stepId":"step//input.js//validStep"}}}}*/;
 // These should all error - only async functions allowed
 export const value = 42;
@@ -11,7 +10,5 @@ export class MyClass {
 export * from './other';
 // This is ok
 export async function validStep() {
-    return __private_run_step("validStep", {
-        arguments: []
-    });
+    return 'allowed';
 }