+++ workbench/nextjs-turbopack/app/api/chat/route.ts
@@ -1,6 +1,6 @@
 // THIS FILE IS JUST FOR TESTING HMR AS AN ENTRY NEEDS
 // TO IMPORT THE WORKFLOWS TO DISCOVER THEM AND WATCH
-import * as workflows from '@/workflows/streams';
+import * as workflows from '@/workflows/3_streams';
 
 export async function POST(_req: Request) {
   console.log(workflows);